refactor down props for router
